Allow Rack apps to handle http OPTIONS requests (TORQUE-792)
Instead of only executing Rack code if all other filters in the chain
return an error, execute Rack code if other filters in the chain
return an error or the response is still not committed. This lets
OPTIONS requests get handled by HttpServlet and Rack applications so
CORS (Cross-Origin Resource Sharing) headers can be set appropriately.