Do a better job of closing -knob.yml MountHandles (TORQUE-570)
The structure of our KnobRootMountProcessor now more closely matches
upstream AS7's DeploymentRootMountProcessor. However, the main bug fix
is attaching our -knob.yml ResourceRoot as a new KNOB_ROOT attachment
on the DeploymentUnit so that we can retrieve it later to unmount. We
can't rely on unmounting the DEPLOYMENT_ROOT attachment because that
gets changed in other processors to point to the actual app's root.

The previous attempted fix for this just kept track of the -knob.yml
MountHandle as a member variable but that would only work if the app
being undeployed was also the last app deployed. Any other case would
lead to the mount not getting unmounted and the error in TORQUE-570.